test(governance): decouple unit and integration tests for ContextSteward
- Move real-file integration tests to tests/integration/
- Ensure tests/unit/ remains isolated and uses mocks
- Comply with TEST-STRUCTURE-STANDARD.md Standard 4 & 7
